cnn/requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
array_record==0.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
chex==0.1.90
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
dm-tree==0.1.9
docstring_parser==0.17.0
einops==0.8.1
etils==1.13.0
executing==2.2.0
flatbuffers==25.2.10
flax==0.11.2
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
h5py==3.14.0
humanize==4.13.0
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jax==0.7.1
jaxlib==0.7.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.11.3
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
msgpack==1.1.1
namex==0.1.0
nest-asyncio==1.6.0
numpy==2.3.2
opt_einsum==3.4.0
optax==0.2.5
optree==0.17.0
orbax-checkpoint==0.11.24
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
promise==2.3
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==27.0.2
requests==2.32.5
rich==14.1.0
scipy==1.16.1
simple-parsing==0.1.7
simplejson==3.20.1
six==1.17.0
stack-data==0.6.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-datasets==4.9.9
tensorflow-metadata==1.17.2
tensorstore==0.1.76
termcolor==3.1.0
toml==0.10.2
toolz==1.0.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
treescope==0.1.10
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0
